print

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -2,8 +2,6 @@
 #include <stdlib.h>
 #include "lists.h"
 
-
-
 /**
  * print_listint - print the elements of all nodes
  * @h: address to head of linked list
@@ -12,7 +10,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-size_t count = 0;
+int count = 0;
 
 while (h != NULL)
 {
